+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## WIP
+
+- Small code speedup in `BaseFader`
+
 ## v2.1.0 (2021-12-16)
 
 - Removed default speeds from methods